(documentation) Add "login" instruction for GitLab
The original steps for GitLab seemed to imply that the Customer would be
setting up a non-admin account to use and even goes so far as to be
clear about group membership. However, once the memberships and
permissions are accomplished, the instructions do not include logging in
as the prepared user. A Customer following the steps may potentially
stay logged in as "admin" in GitLab. While this isn't the worst thing,
it does not follow the overall idea of the instructions.

This commit adds an instruction telling the reader to log back in as the
correct user to complete the steps.